+++ communication/communicators/ClientComm.hpp
@@ -28,7 +28,7 @@ public:
      * Destructor
      */
     ~ClientComm() override = default;
-    void set_timeout_recv(int new_timeout) override;
+    void set_timeout_recv(int64_t new_timeout) override;
     int get_timeout_recv() override;
     virtual bool signon(const utils::Header& header,
 			Comm_t* async_comm=nullptr);
