+++ test/unittest/tests/communicators/commbasetest.cpp
@@ -44,7 +44,7 @@ public:
     handle = new int();
     updateMaxMsgSize(1000);
   }
-  int wait_for_recv(const int&) override { return 0; }
+  int wait_for_recv(const int64_t&) override { return 0; }
   int comm_nmsg(DIRECTION dir=NONE) const override {
     if (nmsg_ >= 0) return nmsg_;
     return CommBase::comm_nmsg(dir);
